vkazhdan 18-add-user-service Migrate Boolean => boolean in User Model. Add checked exceptions to UserService. Add PathVariable to UserRestController
(cherry picked from commit fee0e2855369b069518e96678c581e7ae6764b9d)

+++ cm/cm-biz-local/src/main/java/com/dj/cm/biz/service/echo/impl/UserServiceImpl.java
@@ -1,10 +1,16 @@
 package com.dj.cm.biz.service.echo.impl;
 
+import com.dj.cm.biz.service.exception.AlreadyExistBizExeption;
+import com.dj.cm.biz.service.exception.NotFoundBizException;
 import com.dj.cm.biz.service.user.UserService;
 import com.dj.cm.model.entity.UserModel;
 import com.dj.cm.persistence.repo.user.UserRepository;
+import org.apache.catalina.User;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -23,11 +29,25 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long id) {
+        Optional<UserModel> foundUser = userRepository.findById(id);
+        if (!foundUser.isPresent()){
+            throw new NotFoundBizException("User not found");
+        }
         userRepository.deleteById(id);
     }
 
     @Override
     public UserModel saveUser(UserModel userModel) {
         return userRepository.save(userModel);
+    }
+
+    @Override
+    public UserModel updateUser(UserModel userModel) {
+        UserModel userToUpdate = getUserById(userModel.getId());
+        if (userToUpdate == null){
+            throw new NotFoundBizException("User Not Found");
+        }else
+            BeanUtils.copyProperties(userModel, userToUpdate, "id");
+        return userRepository.save(userToUpdate);
     }
 }
